fix similarity assessors
