+++ examples/3_selecting_similar_datasets/select_similar_datasets_by_knn.py
@@ -2,7 +2,7 @@ from sklearn.model_selection import train_test_split
 
 from meta_automl.data_preparation.datasets_loaders import OpenMLDatasetsLoader
 from meta_automl.data_preparation.meta_features_extractors import PymfeExtractor
-from meta_automl.meta_algorithm.datasets_similarity_assessors import KNNSimilarityAssessor
+from meta_automl.meta_algorithm.datasets_similarity_assessors import KNeighborsBasedSimilarityAssessor
 
 
 def main():
@@ -16,10 +16,10 @@ def main():
     # Split datasets to train (preprocessing) and test (actual meta-algorithm objects).
     x_train, x_test = train_test_split(meta_features, train_size=0.75, random_state=42)
     y_train = x_train.index
-    assessor = KNNSimilarityAssessor({'n_neighbors': 1}, n_best=2)
+    assessor = KNeighborsBasedSimilarityAssessor(n_neighbors=3)
     assessor.fit(x_train, y_train)
     # Get models for the best fitting datasets from train.
-    return x_test.index, assessor.predict(x_test)
+    return x_test.index, assessor.predict(x_test, return_distance=True)
 
 
 if __name__ == '__main__':
